Project restructuring.
