+++ ohlcformer/embeddings/interval_encoding.py
